+++ src/azure-cli-core/azure/cli/core/commands/__init__.py
@@ -36,7 +36,7 @@ logger = azlogging.get_az_logger(__name__)
 
 CONFIRM_PARAM_NAME = 'yes'
 
-BLACKLISTED_MODS = ['context', 'container']
+BLACKLISTED_MODS = ['context', 'container', 'shell']
 
 
 class VersionConstraint(object):