+++ src/command_modules/azure-cli-interactive/azure/cli/command_modules/interactive/_help.py
@@ -3,11 +3,9 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-# pylint: disable=line-too-long
+from azure.cli.core.help_files import helps
 
-from azure.cli.core.commands import register_cli_argument
-
-from azclishell.color_styles import get_options as style_options
-
-register_cli_argument('shell', 'style', options_list=('--style', '-s'),
-                      help='The colors of the shell.', choices=style_options())
+helps['interactive'] = """
+            type: command
+            short-summary: Start the interactive shell for Microsoft Azure CLI (preview)
+            """